refactor: split formatters into per-result-kind package (#59)
formatters.py was 823 LOC with no submodule structure — every consumer
had to scan the whole file to find the formatter for the result kind
they cared about, and there was no test surface below CliRunner.

Split into a formatters/ package with one module per result kind:
trees, evaluation, elasticity, rca, validation, trend, comparison,
investigation, plus a shared _common module for number/date/window
primitives. formatters/__init__.py re-exports the public format_*
functions so existing imports (from qluent_cli.formatters import …)
continue to work unchanged.

formatters/rca.py now consumes RCAOutput from #58 at its interface.

Also refreshed the project-structure section in CLAUDE.md, which had
drifted out of sync with the actual src/qluent_cli/ layout.
